+++ perspective_views/views.py
@@ -27,6 +27,7 @@ import pandas as pd
 def perspective(request):
     event_logs_path = os.path.join(settings.MEDIA_ROOT, "event_logs")
     load_log_succes = False
+    log_information = None
 
 
     #TODO Load the Log Information, else throw/redirect to Log Selection
